--- ts-sdk-site-builder/contracts/sites/walrus_site/deps/sui/object.ts
@@ -1,16 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-/**************************************************************
- * THIS FILE IS GENERATED AND SHOULD NOT BE MANUALLY MODIFIED *
- **************************************************************/
-
-
-/** Sui object identifiers */
-
-import { MoveStruct } from '../../../utils/index.js';
-import { bcs } from '@mysten/sui/bcs';
-const $moduleName = '0x2::object';
-export const UID = new MoveStruct({ name: `${$moduleName}::UID`, fields: {
-        id: bcs.Address
-    } });
