--- ts-sdk-site-builder/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "ts-sdk-site-builder",
-  "module": "index.ts",
-  "type": "module",
-  "devDependencies": {
-    "@types/bun": "latest"
-  },
-  "peerDependencies": {
-    "typescript": "^5"
-  },
-  "dependencies": {
-    "@mysten/codegen": "^0.5.9",
-    "@mysten/sui": "^1.44.0",
-    "@mysten/walrus": "^0.8.3",
-    "effect": "^3.19.1",
-    "yaml": "^2.8.1"
-  }
-}
