--- ts-sdk-site-builder/utils/walrus.ts
@@ -1,8 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-// TODO: include the logic of storing blobs/quilts on Walrus.
-
-function publish_quilt() {
-
-}
