+++ ts-sdk/.prettierrc.json
@@ -1,0 +1,9 @@
+{
+    "printWidth": 100,
+    "trailingComma": "es5",
+    "semi": false,
+    "singleQuote": true,
+    "tabWidth": 4,
+    "bracketSpacing": true,
+    "arrowParens": "always"
+}
