+++ ts-sdk/package.json
@@ -1,0 +1,23 @@
+{
+    "name": "ts-sdk-site-builder",
+    "module": "index.ts",
+    "type": "module",
+    "devDependencies": {
+        "@types/bun": "latest"
+    },
+    "peerDependencies": {
+        "typescript": "^5"
+    },
+    "scripts": {
+        "lint:check": "prettier --check .",
+        "lint:write": "prettier --write .",
+        "codegen": "sui-ts-codegen generate"
+    },
+    "dependencies": {
+        "@mysten/codegen": "^0.5.9",
+        "@mysten/sui": "^1.44.0",
+        "@mysten/walrus": "^0.8.3",
+        "effect": "^3.19.1",
+        "yaml": "^2.8.1"
+    }
+}
