+++ ts-sdk/src/errors.ts
@@ -1,0 +1,2 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
